+++ docs/releases.rst
@@ -2,9 +2,13 @@ Releases
 ========
 Version 0.12
 ------------
+**Added**
+
+* ``ResponseAutoformatMixin``.
+
 **Changed**
 
-* Renamed ``handlers`` module to ``web_handlers``;
+* Renamed ``handlers`` module to ``web_handlers``.
 
 Version 0.11
 ------------